bot/requirements.txt
python-telegram-bot==21.10
firebase-admin==6.6.0
python-dotenv==1.1.0
